Move _0 and _1 to common package
