+++ src/commonMain/kotlin/choliver/nespot/cpu/Utils.kt
@@ -2,9 +2,4 @@ package choliver.nespot.cpu
 
 import choliver.nespot.common.Address
 
-@Suppress("ObjectPropertyName")
-const val _0 = false
-@Suppress("ObjectPropertyName")
-const val _1 = true
-
 fun samePage(a: Address, b: Address) = ((a xor b) and 0xFF00) == 0
