+++ src/jvmTest/kotlin/choliver/nespot/apu/FrameSequencerTest.kt
@@ -3,8 +3,8 @@ package choliver.nespot.apu
 import choliver.nespot.apu.FrameSequencer.Mode.FIVE_STEP
 import choliver.nespot.apu.FrameSequencer.Mode.FOUR_STEP
 import choliver.nespot.apu.FrameSequencer.Ticks
-import choliver.nespot.cpu._0
-import choliver.nespot.cpu._1
+import choliver.nespot.common._0
+import choliver.nespot.common._1
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
