+++ src/jvmTest/kotlin/choliver/nespot/cpu/BranchJumpTest.kt
@@ -1,8 +1,6 @@
 package choliver.nespot.cpu
 
-import choliver.nespot.common.Address
-import choliver.nespot.common.hi
-import choliver.nespot.common.lo
+import choliver.nespot.common.*
 import choliver.nespot.cpu.Cpu.Companion.VECTOR_IRQ
 import choliver.nespot.cpu.Opcode.*
 import org.junit.jupiter.api.Test
