+++ src/jvmTest/kotlin/choliver/nespot/cpu/FlagsTest.kt
@@ -1,5 +1,7 @@
 package choliver.nespot.cpu
 
+import choliver.nespot.common._0
+import choliver.nespot.common._1
 import choliver.nespot.cpu.Opcode.*
 import org.junit.jupiter.api.Test
