+++ src/jvmTest/kotlin/choliver/nespot/ppu/PpuTest.kt
@@ -4,12 +4,7 @@ import choliver.nespot.DOTS_PER_CYCLE
 import choliver.nespot.DOTS_PER_SCANLINE
 import choliver.nespot.SCANLINES_PER_FRAME
 import choliver.nespot.SCREEN_HEIGHT
-import choliver.nespot.common.Address
-import choliver.nespot.common.hi
-import choliver.nespot.common.isBitSet
-import choliver.nespot.common.lo
-import choliver.nespot.cpu._0
-import choliver.nespot.cpu._1
+import choliver.nespot.common.*
 import choliver.nespot.memory.Memory
 import choliver.nespot.nes.VideoSink
 import choliver.nespot.ppu.Ppu.Companion.BASE_PALETTE
